requirements.txt
flask==2.2.3
gunicorn==20.1.0
Werkzeug==2.2.3
flask-cors==3.0.10
google-cloud-storage==2.0.0
